ai-service/requirements.txt
# Core AI / Agent Framework
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0

# Chat UI
chainlit>=1.3.0

# Data Visualization
plotly>=5.24.0

# Google Gemini
google-generativeai>=0.8.0

# Database
sqlalchemy>=2.0.0
pymysql>=1.1.0
psycopg2-binary>=2.9.0

# REST API (standalone mode)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.10.0
